[Refactor] : 편의성을 위해 자료 위치 이동 (#28)
