+++ src/main/java/com/potato/balbambalbam/weaksoundtest/controller/PhonemeController.java
@@ -16,46 +16,13 @@ import java.util.Map;
 @RestController
 public class PhonemeController {
 
-    private final PhonemeService phonemeService;
     private final UserWeakSoundRepository userWeakSoundRepository;
 
-    public PhonemeController (PhonemeService phonemeService, UserWeakSoundRepository userWeakSoundRepository){
-        this.phonemeService = phonemeService;
+    public PhonemeController (UserWeakSoundRepository userWeakSoundRepository){
         this.userWeakSoundRepository = userWeakSoundRepository;
     }
 
-    @PostMapping("/test/finalize")
-    public ResponseEntity<?> finalizeAnalysis(@RequestHeader(value = "userId", required = false) Long userId) {
-        if (userId == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("userId 헤더가 필요합니다."); //401
-        }
-
-        Map<Long, Integer> topPhonemes = null;
-
-        try {
-            topPhonemes = phonemeService.getTopPhonemes(userId);
-
-            if (topPhonemes != null && !topPhonemes.isEmpty()) {
-                System.out.println(userId + ":");
-                topPhonemes.forEach((phonemeId, count) -> {
-                    System.out.println("Phoneme ID : " + phonemeId + ", Count : " + count);
-                    UserWeakSound userWeakSound = new UserWeakSound(userId, phonemeId);
-                    userWeakSoundRepository.save(userWeakSound);
-                });
-            } else {
-                System.out.println("user id " + userId + " 의 취약음 테스트 결과가 없습니다.");
-                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("취약음 테스트 결과가 없습니다."); //404
-            }
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("서버 오류가 발생했습니다."); //500
-        } finally {
-            phonemeService.clearTemporaryData(userId);
-            System.out.println("user id " + userId + " 의 임시 저장소를 삭제했습니다.");
-        }
-        return ResponseEntity.ok(topPhonemes); //200
-    }
-
-    @GetMapping("/test/phonemes")
+    @GetMapping("/phonemes")
     public ResponseEntity<?> getWeakPhonemesByUserId(@RequestHeader(value = "userId", required = false) Long userId) {
         if (userId == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("userId 헤더가 필요합니다."); //401
